+++ src/brazilfi/providers/__init__.py
@@ -1,6 +1,7 @@
+from brazilfi.providers.anbima import ANBIMA
 from brazilfi.providers.b3 import B3
 from brazilfi.providers.bacen import Bacen
 from brazilfi.providers.ibge import IBGE
 from brazilfi.providers.tesouro import TesouroDireto
 
-__all__ = ["B3", "IBGE", "Bacen", "TesouroDireto"]
+__all__ = ["ANBIMA", "B3", "IBGE", "Bacen", "TesouroDireto"]